Weapon classes code separated from IWeapon's script
