+++ Assets/Scripts/AutomaticWeapon.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d1e336a191a8c104caf737d76754a238
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
